+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Refactored example project for easier maintainance
 - Fixed bug: cropping image outputed wrong result image (iOS)
-
+- Fixed bug: picking image on web
 
 ## 0.1.1+3